backend/requirements.txt
# FastAPI Core (Python 3.12)
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
python-multipart>=0.0.18,<1.0.0

# Database
sqlalchemy>=2.0.36,<3.0.0
psycopg2-binary>=2.9.10,<3.0.0
alembic>=1.14.0,<2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
pyotp>=2.9.0,<3.0.0
qrcode[pil]>=8.0,<9.0.0
bcrypt>=4.3.0,<6.0.0

# Rate Limiting
slowapi>=0.1.9,<1.0.0

# Bunker Security (Grado Militar)
pynacl>=1.5.0,<2.0.0
hvac>=2.3.0,<3.0.0
argon2-cffi>=25.1.0,<26.0.0

# Validation & Serialization
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
email-validator>=2.2.0,<3.0.0

# Financial Calculations (Python 3.12)
numpy>=2.0.0,<3.0.0
numpy-financial>=1.0.0,<2.0.0
pandas>=2.2.0,<3.0.0
scipy>=1.14.0,<2.0.0

# Cloud & Storage
boto3>=1.35.0,<2.0.0
python-magic>=0.4.27,<1.0.0

# OCR & Document Processing
pdf2image>=1.17.0,<2.0.0
pytesseract>=0.3.13,<1.0.0

# Redis (Sessions & Cache)
redis>=5.3.0,<6.0.0

# Utilities
python-dotenv>=1.0.1,<2.0.0
httpx>=0.28.0,<1.0.0
tenacity>=9.0.0,<10.0.0

# Email (SendGrid)
sendgrid>=6.11.0,<7.0.0

# AI Analysis (Python 3.12)
anthropic>=0.45.0,<1.0.0
PyPDF2>=3.0.1,<4.0.0

# Monitoring & Observability
prometheus-client>=0.21.0,<1.0.0
opentelemetry-api>=1.28.0,<2.0.0
opentelemetry-sdk>=1.28.0,<2.0.0

# Blockchain (Web3 - Python 3.12)
web3>=7.0.0,<8.0.0
eth-account>=0.13.0,<1.0.0
eth-typing>=5.0.0,<6.0.0
eth-hash[pycryptodome]>=0.7.0,<1.0.0

# Testing
pytest>=8.4.0,<9.0.0
pytest-asyncio>=0.26.0,<1.0.0
pytest-cov>=6.3.0,<7.0.0
apscheduler>=3.10.0
